pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["prabin karki"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
SQLAlchemy = "^1.3.19"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
databases = {extras = ["postgresql"], version = "^0.3.2"}
isort = "^5.5.1"
pre-commit = "^2.7.1"
alembic = "^1.4.2"
mypy = "^0.782"
psycopg2 = "^2.8.6"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
flake8 = "^3.8.3"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
